Add route for viewing single post by entity ID
Could not find a way to use blog.html template for single post, despite
creating a similar GQL, just with a single post. The GQL appeared to
return no record. So stuck with the recommended get-by-id method, but
with this method, I don't know how to pass the data to blog.html in the
same manner the data is passed with the 5 most recent posts. So created
a new template, single_post.html, and modeled it off of blog.html, it
just handles the data slightly differently.